Add new string functions and cross-platform support
This commit significantly expands the library with new BASIC functions and improves cross-platform compatibility.

New Functions:
- String manipulation: INSTR, LCASE$, LEFT$, MID$, RIGHT$, UCASE$
- Screen/cursor: POS, CSRLIN
- Utility: ADD (demonstration function)

Cross-Platform Enhancements:
- Created kbd_util module for Unix keyboard operations (_kbhit, _getch, Sleep)
- Created time_util module for Unix time functions (_getsystime)
- Created screen_util module for cursor position detection (all platforms)
- Updated COMMAND$, INKEY$, SLEEP, DATE_TIME, TIMER to work on macOS/Linux
- Updated test infrastructure to skip terminal-dependent tests when appropriate

Build System Improvements:
- Refactored Windows build to support both GCC and Clang toolchains
- Added Makefile.windows.gcc and Makefile.windows.clang
- Updated Makefile.linux to link math library (-lm)
- Enhanced run_all_tests.sh with better platform detection

Bug Fixes:
- Fixed RND function range calculation to avoid overflow
- Fixed function name inconsistencies (ltrim → ltrim$, rtrim → rtrim$)
- Improved test robustness with isatty() checks and better validation

Documentation:
- Updated CLAUDE.md with complete function list (~30 functions)
- Added platform utility module documentation
- Added platform-specific implementation guidelines

+++ main/src/randomize_rnd.c
@@ -41,8 +41,8 @@ double rnd_F64(double seed) {
   }
 
   if (seed != 0.0) {
-    last_number = (double) rand() / (RAND_MAX + 1);
+    last_number = (double) rand() / ((double) RAND_MAX + 1.0);
   }
-  
+
   return last_number;
 }
